requirements.txt
ofxparse==0.21
pandas==2.2.3

langchain_core==0.2.39
langchain_groq==0.1.9

python-dotenv==1.0.1

# langchain_openai==0.1.23
# openai==1.45.0
# streamlit==1.35.0
# plotly==5.24.0